Alternate display for 3 panels results mode
Add an alternate display for results display in 3 panels mode (need new
css file).
Fix a bug on number of empty lines.
Fix a bug on table format when results tables are empty.
More robust support of LoRa receiver allowing disconnection and
reconnection of the radio on the USB port while the server is running.